% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ems.R
\name{theme_ems}
\alias{theme_ems}
\title{EMS plotting theme}
\usage{
theme_ems(base_size = 12)
}
\arguments{
\item{base_size}{Base font size}
}
\value{
A ggplot2 theme object
}
\description{
A simple ggplot theme used for RedoxRRI visualizations.
}
